Update tính năng lưu trữ
1, Click xóa thì xóa file
2, Tên ngắn shortname
3, Trạng thái upload

+++ blogadmin/luutru/request/xoadulieu.php
@@ -4,11 +4,13 @@ include("../../libs/db_connect.php");
 
 
 $id = $_POST["filename"];
-// $filename = '1514530144_sunflower.jpg';
-$sql= 'DELETE FROM `upload` WHERE filename= "'.$id.'"';
+
+if (unlink('../upload/'.$id)){
+	$sql= 'DELETE FROM `upload` WHERE filename= "'.$id.'"';
+	mysqli_query($con, $sql);
+}
 
 
-$result = mysqli_query($con, $sql);
-// echo $result;
+
 
 ?>
